pyproject.toml
[tool.poetry]
name = "synapses-py"
version = "8.0.0"
description = "A lightweight library for neural networks that runs anywhere"

license = "MIT"

authors = [
    "Dimosthenis Michailidis <mrdimosthenis@hotmail.com>"
]

readme = 'README.md'

repository = "https://github.com/mrdimosthenis/Synapses"
homepage = "https://mrdimosthenis.github.io/Synapses"

keywords = ['neural-network', 'machine-learning', 'deep-learning', 'functional-programming']

[tool.poetry.dependencies]
python = "^3.6,>=3.6.1"
PyFunctional = { version = "^1.4.3" }
